requirements.txt
streamlit
boto3pyngrok